changing the structure to an array because its easier to work with arrays sometimes.
